Extract Builder from BoundedContext.
